% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-fxns.R
\name{arbutus_plot_dist}
\alias{arbutus_plot_dist}
\title{Plot the simulated test statistics versus observed test statistic
(distribution of parameter estimates)}
\usage{
arbutus_plot_dist(
  x,
  stat,
  xlab = NA,
  ylab = "Counts",
  col = NA,
  breaks = NA,
  las = 1,
  lwd = 3,
  legend = TRUE,
  legend.pos = "topright",
  cex = 1,
  ...
)
}
\arguments{
\item{x}{an \code{arbutus} object inherited from
\code{\link{compare_pic_stat}}}

\item{stat}{the name of the test statistic to plot}

\item{xlab}{x-axis label. Defaults to \code{stat}.}

\item{ylab}{y-axis label. Defaults to "Counts".}

\item{col}{colours used in plot. The first colour corresponds to the
histogram of simulated test statistics. The second colour to the line
representing the observed value of the test statistic. If no argument is
supplied, the function uses the default colours.}

\item{breaks}{the number of breaks in the histogram. Defaults to
20.}

\item{las}{plot parameter (see \code{\link{par}}).}

\item{lwd}{plot parameter (see \code{\link{par}}).}

\item{legend}{logical, specifying whether a legend should be included.}

\item{legend.pos}{postiion of the legend (see \code{\link{legend}}).}

\item{cex}{plot parameter for legend (see \code{\link{par}}).}

\item{...}{additional arguments to be passed to plot}
}
\description{
Plots a histogram of the simulated values along a histogram of
the observed values for a single test statistic when a distribution of
estimates is used
}
\details{
This function is designed to give researchers a quick visual
inspection of model adequacy by plotting the simulated distribution of
a test statistic along with the distribution of empirical estimates.

If the empirical distribution of the test statistic lies in the tails of the
distribution of simulated test statistics, the model can be rejected
as inadequate.

The p-values corresponding to this plot can be obtained with the function
\code{\link{pvalue_arbutus}}.

\code{\link{plot.arbutus}} wraps this function and allows results from all
test statistics to be viewed in a single plot.
}
\examples{
## finch data
data(finch)
phy <- finch$phy
data <- finch$data[,1]

## pretend that the tree has already been rescaled
## and that we have a list of rescaled trees
mphy <- lapply(seq_len(50), function(x) {phy})
class(mphy) <- "multiPhylo"
modelad <- arbutus(mphy, data=data)
arbutus_plot_dist(modelad, "m.sig")


\dontrun{
## use diversitree to fit a BM model with MCMC
require(diversitree)
lik <- make.bm(phy, data)
f <- mcmc(lik, x.init=1, nsteps=1000, w=1, print.every=0
modelad.bm <- arbutus(f, sample=100)
arbutus_plot_dist(modelad.bm, "m.sig")
}

}
\seealso{
\code{\link{arbutus}}, \code{\link{compare_pic_stat}},
\code{\link{plot.arbutus}}
}
