% Generated by roxygen2 (4.0.1): do not edit by hand
\name{check_pic_stat}
\alias{check_pic_stat}
\title{Internal function to check test statistics}
\usage{
check_pic_stat(stats)
}
\arguments{
\item{stats}{a named list of test statistics}
}
\value{
named list of test statistics after being checked.
}
\description{
Makes sure test statistics sent to \code{\link{calculate_pic_stat}} are in the correct format
}
\examples{
## produce list of test stats (using default test statistics)
check_pic_stat(default_pic_stat())

## use custom list of test statistics
foo <- function(x) mean(x$pics[,"contrasts"])

my.stats <- list(mean.of.contrasts=foo)
check_pic_stat(my.stats)
}
\seealso{
\code{\link{calculate_pic_stat}}
}
\keyword{internal}

