% Generated by roxygen2 (4.0.1): do not edit by hand
\name{default_pic_stat}
\alias{default_pic_stat}
\title{Internal function for getting default test statistics}
\usage{
default_pic_stat()
}
\value{
a named list of functions
}
\description{
Creates a list of the default test statistics to be used
to asses model adequacy
}
\details{
The following test statistics are produced by this function:
 \enumerate{
  \item{m.sig: }{The mean of the squared contrasts. This is equivalent to the REML estimate of sigsq.}

  \item{c.var: }{The coefficient of variation of the absolute value of the contrasts.}

  \item{s.var: }{The slope of a linear model fit between the contrasts and their expected variances.}

  \item{s.anc: }{The slope of a linear model fit between the contrasts and their inferred ancestral state.}

  \item{s.hgt: }{The slope of a linear model fit between the contrasts and the node height at which they were calculated.}

  \item{d.cdf: }{The D-statistic from a KS test comparing the distribution of the contrasts to a normal distribution  with mean 0 and variance equal to the square root of the squared mean of the contrasts.}
 }
}
\examples{
## get default test statistics
stats <- default_pic_stat()
stats
}
\seealso{
\code{\link{calculate_pic_stat}}, \code{\link{pic_stat_msig}}, \code{\link{pic_stat_cvar}},
\code{\link{pic_stat_svar}}, \code{\link{pic_stat_sasr}}, \code{\link{pic_stat_shgt}},
\code{\link{pic_stat_dcdf}}
}
\keyword{internal}

