% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/as.unit.tree.R
\name{is.unit.tree}
\alias{is.unit.tree}
\title{Check if object is a unit.tree}
\usage{
is.unit.tree(x)
}
\arguments{
\item{x}{object}
}
\value{
logical
}
\description{
Utility function for checking if object is of class 'unit.tree'
}
\examples{
## finch data
data(finch)
phy <- finch$phy
data <- finch$data[,"wingL"]


## using just the given phylogeny
unit.tree.phy <- make_unit_tree(phy, data=data)

is.unit.tree(unit.tree.phy)
}
\seealso{
\code{\link{make_unit_tree}}
}

