% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rescale-fitContinuous.R
\name{make_model_phylo}
\alias{make_model_phylo}
\title{Rescale phylogeny based on fitted model parameters}
\usage{
make_model_phylo(x, ...)
}
\arguments{
\item{x}{an object inherited from a \code{model_info} function}

\item{...}{additional arguments to be passed to \code{make_model_phylo}}
}
\value{
a \code{phylo} object that can be used to from a \code{unit.tree} object
}
\description{
Rescales phylogeny to form a 'unit.tree' using parameters
from fitted model
}
\details{
This is a generic function which rescales the phylogeny based on the model
specific information. While the class and information may differ, the object must include
the 'phylo' object to be rescaled and model specific information such as model type and
parameter values.

To include additional types of models, researchers will need to build a new \code{\link{model_info}}
function for the given model type. The output should be assigned a novel class.
A \code{make_model_phylo.class} object will need to be created to perform the rescaling.
}
\examples{
## finch data
data(finch)
phy <- finch$phy
data <- finch$data[,"wingL"]

\dontrun{
require(geiger)
## fit Brownian motion model
## using geiger's fitContinuous function
fit.bm <- fitContinuous(phy=phy, dat=data, model="BM",
                                 control=list(niter=10))

## get model info using internal arbutus function
info.bm <- model_info(fit.bm)

## rescale phylogeny based on model info
make_model_phylo(info.bm)
}
}
\seealso{
\code{\link{make_unit_tree}}
}

