% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-stats.R
\name{pic_stat_cvar}
\alias{pic_stat_cvar}
\title{Coefficient of variation of contrasts}
\usage{
pic_stat_cvar(unit.tree)
}
\arguments{
\item{unit.tree}{a \code{unit.tree} object}
}
\value{
\code{c.var} the coefficient of variation of the contrasts
}
\description{
Calculates the coefficient of variation on a set of contrasts
}
\details{
This test statistic is used to evaluate whether the model is sufficiently
capturing the variation in rates across the tree. The coefficient of variation is the sd/mean.

The coefficient of variation of the contrasts is included as a default test statistic
in the function \code{\link{calculate_pic_stat}}.

The test statistic computed from a single \code{unit.tree} can be visualized
with the function \code{\link{pic_stat_sasr_plot}}.
}
\examples{
data(finch)
phy <- finch$phy
dat <- finch$data[,"wingL"]
unit.tree <- make_unit_tree(phy, data=dat)

## estimate variance of contrasts
pic_stat_cvar(unit.tree)

## visualization
pic_stat_cvar_plot(unit.tree)

}
\seealso{
\code{\link{calculate_pic_stat}}, \code{\link{default_pic_stat}}, \code{\link{pic_stat_cvar_plot}}
}
