% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pic_stat_cvar_plot}
\alias{pic_stat_cvar_plot}
\title{Plot histogram of contrasts with variance}
\usage{
pic_stat_cvar_plot(x, xlab = "Contrasts", ylab = "Counts", col = NA,
  breaks = NA, las = 1, lwd = 2, ...)
}
\arguments{
\item{x}{a \code{unit.tree} object (see \code{\link{make_unit_tree}}).}

\item{xlab}{x-axis label. Defaults to "Contrasts".}

\item{ylab}{y-axis label. Defaults to "Counts".}

\item{col}{plot colours. The first colour is used for the histogram, the
second designates the mean. If no argument supplied, uses default colours.}

\item{breaks}{number of breaks for histogram. Defaults to the number of
contrats/10.}

\item{las}{plot parameter (see \code{\link{par}}). Defaults to 1.}

\item{lwd}{plot parameter (see \code{\link{par}}). Defaults to 2.}

\item{...}{aditional arguments to be passed to \code{\link{plot}}.}
}
\description{
This is a diagnostic plot for looking at the distribution
of contrasts derived from a rescaled 'unit.tree' object.
}
\details{
This function is just intended to be a quick diagnostic for looking
at the distribution of contrats. The standard deviation (the arrows
in the plot)/mean is equal to the value of \code{\link{pic_stat_cvar}}. It
is a default test statistic in \code{\link{calculate_pic_stat}}.

If the model is adequate the sd should be close to 1.

This function can currently only take a single unit.tree (does not
integrate across unit.trees)
}
\examples{
## finch data
data(finch)
phy <- finch$phy
data <- finch$data[,1]

u <- make_unit_tree(phy, data=data)

pic_stat_cvar_plot(u)
}
\seealso{
\code{\link{make_unit_tree}}, \code{\link{calculate_pic_stat}},
\code{\link{pic_stat_cvar}}
}

