% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summary-stats.R
\name{pic_stat_dcdf}
\alias{pic_stat_dcdf}
\title{Evaluate normality of contrasts}
\usage{
pic_stat_dcdf(unit.tree)
}
\arguments{
\item{unit.tree}{a \code{unit.tree} object}
}
\value{
\code{d.cdf} the D-statistic from a KS-test
}
\description{
Performs Kolmogorov-Smirnov test comparing distribution
of contrasts with that of a normal distribution with mean 0 and variance
equal to the square root of the mean of squared contrasts
}
\details{
This test statistics is used to evaluate whether the assumption of multivariate normailty
is appropriate. If the model which generated the data is the fitted model, we expect the
square root ofthe mean of squared contrasts to be equal to 1. The empirical estimate is used
rather than assume a variance of 1 to reduce the overlap between this test statistic and the
REML estimate of sigma2 (see \code{\link{pic_stat_msig}}). The Kolmogorov-Smirnov (KS) test is a
non-parameteric test which computes the maximum distance \eqn{D} between two cumulative distribution functions.
Running the test multiple times on the same data will produce slightly different values due to the fact
that the null distribution is produced by randomly drawing from a normal distribution.

The KS-D statistic is included as a default test statistic
in the function \code{\link{calculate_pic_stat}}.

The test statistic computed from a single \code{unit.tree} can be visualized
with the function \code{\link{pic_stat_dcdf_plot}}.
}
\examples{
data(finch)
phy <- finch$phy
dat <- finch$data[,"wingL"]
unit.tree <- make_unit_tree(phy, data=dat)

## KS-D statistic
pic_stat_dcdf(unit.tree)

## Visualization
pic_stat_dcdf_plot(unit.tree)
}
\seealso{
\code{\link{calculate_pic_stat}}, \code{\link{default_pic_stat}}, \code{\link[stats]{ks.test}}
}

