% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summary-stats.R
\name{pic_stat_msig}
\alias{pic_stat_msig}
\title{REML estimate of BM parameter}
\usage{
pic_stat_msig(unit.tree)
}
\arguments{
\item{unit.tree}{a \code{unit.tree} object}
}
\value{
\code{m.sig} estimate of sigma2
}
\description{
Calculates the mean of the squared contrasts,
which is equal to the REML estimate of the BM rate parameter.
Used as a test statistic to evaluate model adequacy
}
\details{
This test statistic is used to assess the ability of the model
to adequate estimate the rate of evolution. On datasets simulated on a unit.tree
(see \code{\link{simulate_char_unit}}), the expected value is equal to 1.

The REML estimate of sigma2 is included as a default
test statistic in the function \code{\link{calculate_pic_stat}}.

The test statistic computed from a single \code{unit.tree} can be visualized
with the function \code{\link{pic_stat_msig_plot}}.
}
\examples{
data(finch)
phy <- finch$phy
dat <- finch$data[,"wingL"]
unit.tree <- make_unit_tree(phy, data=dat)

## Estimate sigsq from unit.tree

pic_stat_msig(unit.tree)

## Visualization

pic_stat_msig_plot(unit.tree)
}
\seealso{
\code{\link{calculate_pic_stat}}, \code{\link{default_pic_stat}}, \code{\link{pic_stat_msig_plot}}
}

