% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-fxns.R
\name{pic_stat_msig_plot}
\alias{pic_stat_msig_plot}
\title{Plot histogram of squared contrasts}
\usage{
pic_stat_msig_plot(
  x,
  xlab = "Squared contrasts",
  ylab = "Counts",
  col = NA,
  breaks = NA,
  las = 1,
  lwd = 2,
  ...
)
}
\arguments{
\item{x}{a \code{unit.tree} object (see \code{\link{make_unit_tree}}).}

\item{xlab}{x-axis label. Defaults to "Squared contrasts".}

\item{ylab}{y-axis label. Defaults to "Counts".}

\item{col}{plot colours. The first colour is used for the histogram, the
second designates the mean. If no argument supplied, uses default colours.}

\item{breaks}{number of breaks for histogram. Defaults to the number of
contrats/10.}

\item{las}{plot parameter (see \code{\link{par}}). Defaults to 1.}

\item{lwd}{plot parameter (see \code{\link{par}}). Defaults to 2.}

\item{...}{aditional arguments to be passed to \code{\link{plot}}.}
}
\description{
This is a diagnostic plot for looking at the distribution
of squared contrasts derived from a rescaled 'unit.tree' object.
}
\details{
This function is just intended to be a quick diagnostic for looking
at the distribution of squared contrats. The mean value (the vertical line
in the plot) is equal to the value of \code{\link{pic_stat_msig}} and is
the REML estimate of the Brownian motion rate parameter. It is a default
test statistic in \code{\link{calculate_pic_stat}}.

If the model is adequate the mean should be close to 1.

This function can currently only take a single unit.tree (does not
integrate across unit.trees)
}
\examples{
## finch data
data(finch)
phy <- finch$phy
data <- finch$data[,1]

u <- make_unit_tree(phy, data=data)

pic_stat_msig_plot(u)

}
\seealso{
\code{\link{make_unit_tree}}, \code{\link{calculate_pic_stat}},
\code{\link{pic_stat_msig}}
}
