% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-stats.R
\name{pic_stat_sasr}
\alias{pic_stat_sasr}
\title{Relationship between contrasts and inferred ancestral state}
\usage{
pic_stat_sasr(unit.tree)
}
\arguments{
\item{unit.tree}{a \code{unit.tree} object}
}
\value{
\code{s.asr} the estimated slope parameter
}
\description{
Fits a linear model between the absolute value of the contrasts
and the ancestral state inferred at the node at which the contrast was calculated
}
\details{
This test statistic is the estimate of the slope from fitting
a linear model between the contrasts and the inferred ancestral state at the node
at which the contrasts were calculated. It is used
to evaluate whether the model is adequately capturing variation relative to ancestral state.
The ancestral states are calculated using the contrast based approach and are not technically
ancestral state estimates (see Felsenstein 1985 -- American Naturalist); this is conceptually and statistically
different from the maximum likelihood approach to reconstructing ancestral states.
The contrast based approach is
used here as there is a clear expectation of the relationship between the the contrasts and
trait value. If the model is adequate, we expect there not to be a relationship
between the contrasts and their node height (i.e. slope ~ 0).

The slope of the contrasts and their variances is included as a default test statistic
in the function \code{\link{calculate_pic_stat}}.

The test statistic computed from a single \code{unit.tree} can be visualized
with the function \code{\link{pic_stat_sasr_plot}}.
}
\examples{
data(finch)
phy <- finch$phy
dat <- finch$data[,"wingL"]
unit.tree <- make_unit_tree(phy, data=dat)

## estimate slope
pic_stat_sasr(unit.tree)

## visualization
pic_stat_sasr_plot(unit.tree)

}
\seealso{
\code{\link{calculate_pic_stat}}, \code{\link{default_pic_stat}}, \code{\link{pic_stat_sasr_plot}},
\code{\link[stats]{lm}}, \code{\link[ape]{ace}}
}
