% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pic_stat_shgt}
\alias{pic_stat_shgt}
\title{Relationship between contrasts and node height}
\usage{
pic_stat_shgt(unit.tree)
}
\arguments{
\item{unit.tree}{a \code{unit.tree} object}
}
\value{
\code{s.hgt} the estimated slope parameter
}
\description{
Fits a linear model between the absolute value of the contrasts
and the node height at which they were calculated
}
\details{
This test statistic is the estimate of the slope from fitting
a linear model between the contrasts and the node height (i.e. tree depth)
at which they were calculated. It is used
to evaluate whether the model is adequately capturing variation relative to tree depth.
If the model is adequate, we expect there not to be a relationship
between the contrasts and their node height (i.e. slope ~ 0). Note that the node heights
being used here are the absolute node ages from the original phylogeny and not the adjusted
node ages obtained from computing contrasts (see \code{\link{pic}}).

The slope of the contrasts and their variances is included as a default test statistic
in the function \code{\link{calculate_pic_stat}}.

The test statistic computed from a single \code{unit.tree} can be visualized
with the function \code{\link{pic_stat_svar_plot}}.
}
\examples{
data(finch)
phy <- finch$phy
dat <- finch$data[,"wingL"]
unit.tree <- make_unit_tree(phy, data=dat)

## estimate slope
pic_stat_shgt(unit.tree)

## visualizaiton
pic_stat_shgt_plot(unit.tree)
}
\seealso{
\code{\link{calculate_pic_stat}}, \code{\link{default_pic_stat}}, \code{\link{pic_stat_svar_plot}}, \code{\link[stats]{lm}}
}

