% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-stats.R
\name{pic_stat_svar}
\alias{pic_stat_svar}
\title{Relationship between contrasts and their variances}
\usage{
pic_stat_svar(unit.tree)
}
\arguments{
\item{unit.tree}{a \code{unit.tree} object}
}
\value{
\code{s.var} the estimated slope paramter
}
\description{
Fits a linear model between the absolute value of constrasts
and their expected variances
}
\details{
This test statistic estimates the slope from fitting
a linear model between the contrasts and their expected variance. It is used
to evaluate whether the model is adequately capturing variation relative to branch
lengths. If the model is adequate, we expect there not to be a relationship
between the contrasts and their variances (i.e. slope ~ 0).

The slope of the contrasts and their variances is included as a default test statistic
in the function \code{\link{calculate_pic_stat}}

The test statistic computed from a single \code{unit.tree} can be visualized
with the function \code{\link{pic_stat_svar_plot}}.
}
\examples{
data(finch)
phy <- finch$phy
dat <- finch$data[,"wingL"]
unit.tree <- make_unit_tree(phy, data=dat)

## estimate slope
pic_stat_svar(unit.tree)

## visualization
pic_stat_svar_plot(unit.tree)

}
\seealso{
\code{\link{calculate_pic_stat}}, \code{\link{default_pic_stat}},
\code{\link{pic_stat_svar_plot}}, \code{\link[stats]{lm}}
}
