% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pic_stat_svar_plot}
\alias{pic_stat_svar_plot}
\title{Plot contrasts versus their variances}
\usage{
pic_stat_svar_plot(x, xlab = "Variance", ylab = "Contrast", las = 1,
  col = NA, pch = 19, cex = 1, lwd = 2, ...)
}
\arguments{
\item{x}{a \code{unit.tree} object (see \code{\link{make_unit_tree}}).}

\item{xlab}{x-axis label. Defaults to "Variance".}

\item{ylab}{y-axis label. Defaults to "Contrast".}

\item{las}{plot parameter (see \code{\link{par}}). Defaults to 1.}

\item{col}{plot colours. The first colour is the colour of the points, the second, the colour of the regression line. If no argument, uses default colours.}

\item{pch}{plot parameter (see \code{\link{par}}). Defaults to 19.}

\item{cex}{plot parameter (see \code{\link{par}}). Defaults to 1.}

\item{lwd}{plot parameter (see \code{\link{par}}). Defaults to 2.}

\item{...}{additional arguments to be passed to \code{\link{plot}}.}
}
\description{
Plot the contrasts derived from a rescaled 'unit.tree' against
their expected variances.
}
\details{
This function generates a plot of the observed contrasts versus their
expected variances (see \code{\link[ape]{pic}}).
The slope of this plot is equal to
the test statistic \code{\link{pic_stat_svar}}
and is used as a default test statistic in
\code{\link{calculate_pic_stat}}.

If the model is a good fit to the data (i.e. an adequate model),
the slope ~ 0. The purpose of the plot is to get a quick visual glance
at the data and to assess whether or not the model inadequacy may be driven by outliers.

The function can currently only take a single unit.tree
(does not integrate across multiple trees).
}
\examples{
## finch data
data(finch)
phy <- finch$phy
data <- finch$data[,"wingL"]

u <- make_unit_tree(phy, data=data)

pic_stat_svar_plot(u)
}
\seealso{
\code{\link{make_unit_tree}}, \code{\link{calculate_pic_stat}},
\code{\link{pic_stat_svar}}
}

