% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.arbutus}
\alias{plot.arbutus}
\title{Plot the simulated test statistics versus observed test statistics}
\usage{
\method{plot}{arbutus}(x, xlab = NA, ylab = "Counts", col = NA,
  breaks = NA, las = 1, lwd = 3, legend = TRUE,
  legend.pos = "topright", cex = 1, n.row = 2, n.col = NA, mar = c(5,
  4, 2, 2), ...)
}
\arguments{
\item{x}{an \code{arbutus} object inherited from
\code{\link{compare_pic_stat}}}

\item{xlab}{x-axis label. Defaults to \code{stat}.}

\item{ylab}{y-axis label. Defaults to "Counts".}

\item{col}{colours used in plot. The first colour corresponds to the
histogram of simulated test statistics. The second colour to the line, or
histogram,
representing the observed value of the test statistic. If no argument is
supplied, the function uses the default colours.}

\item{breaks}{the number of breaks in the histogram. Defaults to
20.}

\item{las}{plot parameter (see \code{\link{par}}).}

\item{lwd}{plot parameter (see \code{\link{par}}).}

\item{legend}{logical, specifying whether a legend should be included (see
details).}

\item{legend.pos}{postiion of the legend (see \code{\link{legend}}).}

\item{cex}{plot parameter for legend (see \code{\link{par}}).}

\item{n.row}{number of rows to plot when plotting multiple test statistics.
Defaults to 2.}

\item{n.col}{number of columns to plot when plotting multiple test statistics.
Defaults to \code{ncol(obs)/2}.}

\item{mar}{margins for each plot. Defaults to 5,4,2,2.}

\item{...}{additional arguments to be passed to plot}
}
\description{
Plots a histogram of the simulated values along with
the observed values for a single test statistic. Either a point estimate or
distribution of observed values may be used. Is the generic \code{\link{plot}}
function for objects of class \code{arbutus}.
}
\details{
This function is designed to give researchers a quick visual
inspection of model adequacy by plotting the simulated distribution of
all test statistics versus the emprical values.

This function is a simply a wrapper for the functions
\code{\link{arbutus_plot_point}} and \code{\link{arbutus_plot_dist}}. It
creates plots for all test statistics used in \code{\link{arbutus}} and
\code{\link{compare_pic_stat}}. The function can be supplied either a point
estimate for the observed test statistics or a distribution of estimates.

If the empirical estimate lies in the tails of the distribution of simulated
test statistics, the model can be rejected as inadequate.

The p-values corresponding to this plot can be obtained with the function
\code{\link{pvalue_arbutus}}.

We are still working on a general function for plotting legends. For the time
being, if \code{legend=TRUE}, a legend will appear with each test statistic.
This is obviously not ideal, but we are working on it.

If users wish to create their own plots, all of the information is stored
in the \code{arbutus} object. \code{x$obs} contains a data.frame of observed
test statistics and \code{x$sim}, a data.frame of simulated test statistics.
}
\examples{
## finch data
data(finch)
phy <- finch$phy
data <- finch$data[,1]

## pretend that the tree has already been rescaled
modelad <- arbutus(phy, data=data, nsim=20)
plot(modelad)


\dontrun{
## use diversitree to fit a BM model with ML
require(diversitree)
lik <- make.bm(phy, data)
f <- find.mle(lik, x.init=1)
modelad.bm.ml <- arbutus(f)
plot(modelad.bm.ml, legend=FALSE)

## use diversitree to fit a BM model with MCMC
m <- mcmc(lik, x.init=1, nsteps=1000, w=1, print.every=0)
modelad.bm.mcmc <- arbutus(m, sample=100)
plot(modelad.bm.mcmc)
}
}
\seealso{
\code{\link{arbutus}}, \code{\link{compare_pic_stat}},
\code{\link{arbutus_plot_point}}, \code{\link{arbutus_plot_dist}}
}

