% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pvalue_arbutus}
\alias{pvalue_arbutus}
\title{Extract p--values for test statistics}
\usage{
pvalue_arbutus(x)
}
\arguments{
\item{x}{an \code{arbutus} object from the function \code{\link{compare_pic_stat}}}
}
\value{
a named vector of two-tailed p-values
}
\description{
Utility function for extracting p-values from the output
of \code{\link{compare_pic_stat}}
}
\examples{
data(finch)
phy <- finch$phy
dat <- finch$data[,"wingL"]
unit.tree <- make_unit_tree(phy, data=dat)

## calculate default test stats on observed data
obs <- calculate_pic_stat(unit.tree, stats=NULL)

## simulate data on unit.tree
sim.dat <- simulate_char_unit(unit.tree, nsim=10)

## calculate default test stats on simulated data
sim <- calculate_pic_stat(sim.dat, stats=NULL)

## compare simulated to observed test statistics
res <- compare_pic_stat(obs, sim)

## get p-values
pvalue_arbutus(res)

## note these are returned by default with print.arbutus
res
}
\seealso{
\code{\link{compare_pic_stat}}
}

