\docType{data}
\name{finch}
\alias{finch}
\title{Galapagos finch dataset}
\format{a list with two elements}
\description{
  An example dataset including a phylogeny of 13 Galapagos
  finches with trait data matched to the phylogeny.
}
\details{
  \describe{ \item{phy}{phylogeny of Galapagos finches}
  \item{data}{measured traits for each species in the
  phylogeny} }
}
\keyword{datasets}

