\docType{data}
\name{finch}
\alias{finch}
\title{Galapagos finch dataset}
\format{a list with two elements}
\description{
  A dataset including a phylogeny of 13 galapagos finches
  with trait data matched to the phylogeny.
}
\details{
  \itemize{ \item phy. the finch phylogeny \item data.
  measured traits for each species in the phylogeny }
}
\keyword{datasets}

